// frontend/src/app/models/Education.ts
export class Education {
    _id: string;
    edu_type: string;
    name: string;
    city: string;
    state: string;
    country: string;
    pincode: string;
    qualification: string;
    area_of_study: string;
    web_page: string;
    period_from: Date;
    period_to: Date;
    description: string;
}